Inline `create_pointer()` method from `PointerTensor` to `TorchTensor`
It's only used in one place and is actually about creating
`PointerTensors` from `TorchTensors` anyway, so it rightfully belongs on
`TorchTensor`, not `PointerTensor`.
