+++ syft/frameworks/torch/tensors/interpreters/native.py
@@ -575,11 +575,40 @@ class TorchTensor(AbstractTensor):
         shape=None,
         **kwargs,
     ) -> PointerTensor:
-        """Creates a pointer to the "self" torch.Tensor object.
+        """Creates a pointer to the "self" FrameworkTensor object.
+
+        This method is called on a FrameworkTensor object, returning a pointer
+        to that object. This method is the CORRECT way to create a pointer,
+        and the parameters of this method give all possible attributes that
+        a pointer can be created with.
+
+        Args:
+            location: The AbstractWorker object which points to the worker on which
+                this pointer's object can be found. In nearly all cases, this
+                is self.owner and so this attribute can usually be left blank.
+                Very rarely you may know that you are about to move the Tensor
+                to another worker so you can pre-initialize the location
+                attribute of the pointer to some other worker, but this is a
+                rare exception.
+            id_at_location: A string or integer id of the tensor being pointed
+                to. Similar to location, this parameter is almost always
+                self.id and so you can leave this parameter to None. The only
+                exception is if you happen to know that the ID is going to be
+                something different than self.id, but again this is very rare
+                and most of the time, setting this means that you are probably
+                doing something you shouldn't.
+            owner: A AbstractWorker parameter to specify the worker on which the
+                pointer is located. It is also where the pointer is registered
+                if register is set to True.
+            ptr_id: A string or integer parameter to specify the id of the pointer
+                in case you wish to set it manually for any special reason.
+                Otherwise, it will be set randomly.
+            garbage_collect_data: If true (default), delete the remote tensor when the
+                pointer is deleted.
 
         Returns:
-            A PointerTensor pointer to self. Note that this
-            object will likely be wrapped by a torch.Tensor wrapper.
+            A FrameworkTensor[PointerTensor] pointer to self. Note that this
+            object itself will likely be wrapped by a FrameworkTensor wrapper.
         """
         if id_at_location is None:
             id_at_location = self.id
@@ -593,16 +622,21 @@ class TorchTensor(AbstractTensor):
         if shape is None:
             shape = self.shape
 
-        # ptr = syft.PointerTensor(
-        #     location=location,
-        #     id_at_location=id_at_location,
-        #     owner=owner,
-        #     id=ptr_id,
-        #     garbage_collect_data=garbage_collect_data,
-        #     shape=shape)
+        if owner is None:
+            owner = self.owner
 
-        ptr = syft.PointerTensor.create_pointer(
-            self, location, id_at_location, owner, ptr_id, garbage_collect_data, shape
+        if location is None:
+            location = self.owner
+
+        ptr = PointerTensor(
+            location=location,
+            id_at_location=id_at_location,
+            owner=owner,
+            id=ptr_id,
+            garbage_collect_data=True if garbage_collect_data is None else garbage_collect_data,
+            shape=shape,
+            tags=self.tags,
+            description=self.description,
         )
 
         return ptr
